Feat: 채팅 모듈 테스트 파일 구현

+++ src/providers/database/mongoose/mongoose.seed.ts
@@ -3,24 +3,30 @@ import { ChatRoomModel } from './chatRoom.schema';
 import { ChatModel } from './chat.schema';
 import CHAT_ROOMS from './mock/chatRoom.mock';
 import 'dotenv/config';
+import CHATS from './mock/chat.mock';
 
 async function connectDB() {
   await mongoose.connect(process.env.MONGO_URI);
 }
-
-export async function seed() {
+let isSeeded = false;
+export async function mongooseSeed() {
+  if (isSeeded) {
+    return;
+  }
   await connectDB();
 
   await ChatModel.deleteMany();
   await ChatRoomModel.deleteMany();
 
-  await ChatRoomModel.insertMany(CHAT_ROOMS);
+  const chatRooms = await ChatRoomModel.insertMany(CHAT_ROOMS);
+  await ChatModel.insertMany(CHATS);
 
   console.log('🌱 Mongoose Seeding completed!');
+  isSeeded = true;
   mongoose.connection.close();
 }
 
-seed().catch((e) => {
+mongooseSeed().catch((e) => {
   console.error(e);
   mongoose.connection.close();
 });
